# specs/001-token-reward-and-goal-task-hierarchy/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Goal Together Hierarchy and Rewards API
  version: 1.0.0
  description: API contract for main goal/sub goal/task hierarchy with token rewards.
servers:
  - url: http://localhost:8000
paths:
...
  /sub-goals/{subGoalId}/tasks:
    get:
      summary: List tasks under sub goal
      parameters:
        - in: path
          name: subGoalId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task list
    post:
      summary: Create draft task under sub goal
      parameters:
        - in: path
          name: subGoalId
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Draft task created
...
  /tasks/{taskId}:
    patch:
      summary: Edit task (allowed only in draft)
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task updated
        "409":
          description: Task is confirmed and immutable for edits
    delete:
      summary: Delete task (allowed only in draft)
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Draft task deleted
        "409":
          description: Confirmed task deletion forbidden
...